dns-resolver/requirements.txt
asyncio
pycares
asyncio-nats-client

